NewsPaper/requirements.txt
Django==4.2.7
django-filter==23.3